remove export base file, add export sandbox that hardcodes cuda version
